admin system, disallow users from modifying other users data

+++ src/controllers/userController.js
@@ -17,14 +17,19 @@ export const createUser = (req, res) => {
 };
 
 export const listUsers = (req, res) => {
-    User.find({})
-    .exec((err, users) => {
-        if(err) {
-            res.status(400).send(err);
-        } else {
-            res.status(200).json(users)
-        }
-    });
+    if(req.decoded.data.isAdmin === true){
+        User.find({})
+        .exec((err, users) => {
+            if(err) {
+                res.status(400).send(err);
+            } else {
+                res.status(200).json(users)
+            }
+        })
+    }
+    else{
+        res.status(403)
+    }
 };
 
 export const getUser = (req, res) => {
@@ -62,7 +67,7 @@ export const login = (req, res) => {
         else{
             bcrypt.compare(req.body.password, user.password, function(err, response) {
                 if(response) {
-                    let token = jwt.sign({exp: Math.floor(Date.now() / 1000) + (60 * 60), data: {id: user.id}}, 'mySuperSecrett');
+                    let token = jwt.sign({exp: Math.floor(Date.now() / 1000) + (60 * 60), data: {id: user.id, isAdmin: user.isAdmin}}, 'mySuperSecrett');
                     const response = {user: user, token: token}
                     res.status(200).json(response)
                 } else {
@@ -75,32 +80,42 @@ export const login = (req, res) => {
 };
 
 export const updateUser = (req, res) => {
-    User.findOneAndUpdate({"_id": req.params.id}, req.body, {new: true, useFindAndModify: false})
-    .exec((err, user) => {
-        if(err) {
-            res.status(400).send(err);
-        } else {
-            if(user == null) {
-                res.sendStatus(404);
+    if(req.decoded.data.id === req.params.id || req.decoded.data.isAdmin === true){
+        User.findOneAndUpdate({"_id": req.params.id}, req.body, {new: true, useFindAndModify: false})
+        .exec((err, user) => {
+            if(err) {
+                res.status(400).send(err);
+            } else {
+                if(user == null) {
+                    res.sendStatus(404);
+                }
+                else {
+                    res.status(200).json(user);
+                }
             }
-            else {
-                res.status(200).json(user);
-            }
-        }
-    });
+        });
+    }
+    else{
+        res.sendStatus(403)
+    }
 };
 
 export const deleteUser = (req, res) => {
-    User.findOneAndDelete({"_id": req.params.id}, (err, user) => {
-        if(err) {
-            res.status(400).send(err);
-        } else {
-            if(user == null) {
-                res.sendStatus(404);
+    if(req.decoded.data.id === req.params.id || req.decoded.data.isAdmin === true){
+        User.findOneAndDelete({"_id": req.params.id}, (err, user) => {
+            if(err) {
+                res.status(400).send(err);
+            } else {
+                if(user == null) {
+                    res.sendStatus(404);
+                }
+                else {
+                    res.sendStatus(204);
+                }
             }
-            else {
-                res.sendStatus(204);
-            }
-        }
-    });
+        });
+    }
+    else{
+        res.sendStatus(403)
+    }
 };
